Creating a new testing structure
This diff does a couple things:
- It creates test file for every Model and Document method that we'll need to handle and test. These can be gradually filled in.
- It breaks helpers.js functions into separate files. Each one of those now has a corresponding test file.

In upcoming diffs, I'll move the existing tests into this new structure.
