+++ __tests__/authIsDisabled.test.js
@@ -1,0 +1,3 @@
+const test = require('ava');
+
+test.todo('Create tests for authIsDisabled');
