+++ __tests__/getAuthorizedFields.test.js
@@ -1,0 +1,3 @@
+const test = require('ava');
+
+test.todo('Migrate tests for getAuthorizedFields');
