+++ __tests__/getUpdatePaths.test.js
@@ -1,0 +1,3 @@
+const test = require('ava');
+
+test.todo('Migrate getUpdatePaths tests to this file');
