+++ __tests__/hasPermission.test.js
@@ -1,0 +1,3 @@
+const test = require('ava');
+
+test.todo('Migrate hasPermission tests here');
