+++ __tests__/methods/Document.save.test.js
@@ -1,0 +1,3 @@
+const test = require('ava');
+
+test.todo('Write tests for Document.save');
