+++ __tests__/methods/Model.aggreate.test.js
@@ -1,0 +1,3 @@
+const test = require('ava');
+
+test.todo('Write tests for Model.aggregate. It should be a disabled method');
