+++ __tests__/methods/Model.bulkWrite.test.js
@@ -1,0 +1,3 @@
+const test = require('ava');
+
+test.todo('Write tests for Model.bulkWrite. It should be disabled, does not support middleware');
