+++ __tests__/methods/Model.distinct.test.js
@@ -1,0 +1,3 @@
+const test = require('ava');
+
+test.todo('Write tests for Model.distinct');
