+++ __tests__/resolveAuthLevel.test.js
@@ -1,0 +1,3 @@
+const test = require('ava');
+
+test.todo('Migrate resolveAuthLevel tests to here');
