+++ __tests__/sanitizeDocumentList.test.js
@@ -1,0 +1,3 @@
+const test = require('ava');
+
+test.todo('Migrate SanitizeDocumentList tests to this file');
