+++ src/PermissionDeniedError.js
