+++ src/authIsDisabled.js
@@ -1,0 +1,5 @@
+function authIsDisabled(options) {
+  return options && options.authLevel === false;
+}
+
+module.exports = authIsDisabled;
